+++ file/example_test.go
@@ -37,7 +37,7 @@ func ExampleNew() {
 		MaxBackups:  5,
 		MaxAgeDays:  30,
 		Permissions: "0600",
-	}, nil)
+	})
 	if err != nil {
 		fmt.Println("create error:", err)
 		return